[chore][pkg/stanza] Cleanup output operator files
