+++ pkg/stanza/operator/output/drop/output_test.go
