+++ pkg/stanza/operator/output/file/output.go
@@ -1,0 +1,71 @@
+// Copyright The OpenTelemetry Authors
+// SPDX-License-Identifier: Apache-2.0
+
+package file // import "github.com/open-telemetry/opentelemetry-collector-contrib/pkg/stanza/operator/output/file"
+
+import (
+	"context"
+	"encoding/json"
+	"html/template"
+	"os"
+	"sync"
+
+	"github.com/open-telemetry/opentelemetry-collector-contrib/pkg/stanza/entry"
+	"github.com/open-telemetry/opentelemetry-collector-contrib/pkg/stanza/operator"
+	"github.com/open-telemetry/opentelemetry-collector-contrib/pkg/stanza/operator/helper"
+)
+
+// Output is an operator that writes logs to a file.
+type Output struct {
+	helper.OutputOperator
+
+	path    string
+	tmpl    *template.Template
+	encoder *json.Encoder
+	file    *os.File
+	mux     sync.Mutex
+}
+
+// Start will open the output file.
+func (o *Output) Start(_ operator.Persister) error {
+	var err error
+	o.file, err = os.OpenFile(o.path, os.O_RDWR|os.O_APPEND|os.O_CREATE, 0600)
+	if err != nil {
+		return err
+	}
+
+	o.encoder = json.NewEncoder(o.file)
+	o.encoder.SetEscapeHTML(false)
+
+	return nil
+}
+
+// Stop will close the output file.
+func (o *Output) Stop() error {
+	if o.file != nil {
+		if err := o.file.Close(); err != nil {
+			o.Errorf(err.Error())
+		}
+	}
+	return nil
+}
+
+// Process will write an entry to the output file.
+func (o *Output) Process(_ context.Context, entry *entry.Entry) error {
+	o.mux.Lock()
+	defer o.mux.Unlock()
+
+	if o.tmpl != nil {
+		err := o.tmpl.Execute(o.file, entry)
+		if err != nil {
+			return err
+		}
+	} else {
+		err := o.encoder.Encode(entry)
+		if err != nil {
+			return err
+		}
+	}
+
+	return nil
+}
