+++ pkg/stanza/operator/output/stdout/config.go
@@ -4,30 +4,29 @@
 package stdout // import "github.com/open-telemetry/opentelemetry-collector-contrib/pkg/stanza/operator/output/stdout"
 
 import (
-	"context"
 	"encoding/json"
 	"io"
 	"os"
-	"sync"
 
 	"go.uber.org/zap"
 
-	"github.com/open-telemetry/opentelemetry-collector-contrib/pkg/stanza/entry"
 	"github.com/open-telemetry/opentelemetry-collector-contrib/pkg/stanza/operator"
 	"github.com/open-telemetry/opentelemetry-collector-contrib/pkg/stanza/operator/helper"
 )
+
+const operatorType = "stdout"
 
 // Stdout is a global handle to standard output
 var Stdout io.Writer = os.Stdout
 
 func init() {
-	operator.Register("stdout", func() operator.Builder { return NewConfig("") })
+	operator.Register(operatorType, func() operator.Builder { return NewConfig("") })
 }
 
 // NewConfig creates a new stdout config with default values
 func NewConfig(operatorID string) *Config {
 	return &Config{
-		OutputConfig: helper.NewOutputConfig(operatorID, "stdout"),
+		OutputConfig: helper.NewOutputConfig(operatorID, operatorType),
 	}
 }
 
@@ -47,24 +46,4 @@ func (c Config) Build(logger *zap.SugaredLogger) (operator.Operator, error) {
 		OutputOperator: outputOperator,
 		encoder:        json.NewEncoder(Stdout),
 	}, nil
-}
-
-// Output is an operator that logs entries using stdout.
-type Output struct {
-	helper.OutputOperator
-	encoder *json.Encoder
-	mux     sync.Mutex
-}
-
-// Process will log entries received.
-func (o *Output) Process(_ context.Context, entry *entry.Entry) error {
-	o.mux.Lock()
-	err := o.encoder.Encode(entry)
-	if err != nil {
-		o.mux.Unlock()
-		o.Errorf("Failed to process entry: %s", err)
-		return err
-	}
-	o.mux.Unlock()
-	return nil
 }
