+++ pkg/stanza/operator/output/stdout/output_test.go
